♻️ refactor: split WorkModel god-class into per-type work modules
Structurally move the 1433-line monolithic WorkModel into a work/ directory,
splitting the polymorphic Work registry per provider type (task / document /
linear / github) without changing any behavior or the public API.

- context.ts: WorkContext + ownership free functions (replaces `this`)
- internal.ts / cost.ts / writes.ts: shared leaves + cross-type write pipeline
- task/document/linear/github.ts: per-type register + list functions
- queries.ts: cross-type read orchestration
- index.ts: WorkModel facade delegating each public method to a free function

Per-type modules take an explicit `ctx` instead of `this`, so they never
depend back on the facade class (acyclic module graph). work.test.ts is
untouched and all 26 cases pass.
